tags crud, add delete user validation

+++ src/users/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Put, Param, Delete, HttpCode, UseGuards } from '@nestjs/common';
+import { Controller, Get, Request, Body, Put, Param, Delete, HttpCode, UseGuards, UnauthorizedException, ForbiddenException } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
@@ -19,13 +19,17 @@ export class UserController {
   }
 
   @Put(':id')
+  @HttpCode(204)
   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
     return this.userService.update(+id, updateUserDto);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @Request() req) {
+    if (req.user.id === +id) {
+      throw new ForbiddenException("Cannot delete yourself.");
+    }
     return this.userService.remove(+id);
   }
 }
